pyproject.toml
[project]
name = "pygame_manager"
version = "1.0.0"
authors = [{ name = "AntonisPylos", email = "antonis@pylos.dev" }]
description = "A CLI project management library for pygame community edition."
keywords = ["pygame", "cli", "project manager"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "appdirs==1.4.4",
    "cx_Freeze == 8.5.3",
    "GitPython == 3.1.46"
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: pygame",
]
[project.urls]
Repository = "https://github.com/AntonisPylos/pygame-manager"
Issues = "https://github.com/AntonisPylos/pygame-manager/issues"
Releases = "https://github.com/AntonisPylos/pygame-manager/releases"

[project.scripts]
pygame = "pygame_manager.__main__:main"
pygame-ce = "pygame_manager.__main__:main"
pgce = "pygame_manager.__main__:main"

[tool.setuptools]
package-dir = { "pygame_manager" = "src" }
packages = [
    "pygame_manager"
]

[tool.setuptools.package-data]
pygame_manager = ["template/**"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
